unify tintDrawable methods
